Introduce AppUpdater
