+++ app/src/main/res/values-zh-rCN/strings.xml
@@ -511,9 +511,7 @@
     <string name="account_name">账户</string>
     <string name="preload_thumb_aggressively">激进地预载缩略图</string>
     <string name="new_version_available">有新版本！</string>
-    <string name="ci_is_running">CI 正在运行，请稍候</string>
-    <string name="already_latest_ci_build">已是最新 CI 构建</string>
-    <string name="already_latest_release">已是最新版本 %s</string>
+    <string name="already_latest_version">已是最新版本</string>
     <!-- Tachiyomi Reader Start-->
     <string name="action_settings">设置</string>
     <string name="action_retry">重试</string>
