+++ app/src/main/res/values/strings.xml
@@ -574,9 +574,7 @@
     <string name="account_name">Account</string>
     <string name="preload_thumb_aggressively">Preload thumbs aggressively</string>
     <string name="new_version_available">New version available!</string>
-    <string name="ci_is_running">CI is still running, please wait</string>
-    <string name="already_latest_ci_build">Already latest CI build</string>
-    <string name="already_latest_release">Already latest Release %s</string>
+    <string name="already_latest_version">Already the latest version</string>
     <!-- Tachiyomi Reader Start-->
     <string name="action_settings">Settings</string>
     <string name="pref_rotation_type">Default rotation type</string>
